% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotWithDots.R
\name{boxplotWithDots}
\alias{boxplotWithDots}
\title{boxplotWithDots}
\usage{
boxplotWithDots(myExpres, lev, aTitle, groupLabels, addBoxplot = TRUE)
}
\arguments{
\item{myExpres}{Numeric values}

\item{lev}{A factor for drawing the plot as plot(expres~lev). Pay attention to the order of values in the expres vector and levels in the factor}

\item{aTitle}{The title for the plot}

\item{groupLabels}{Names for the groups}

\item{addBoxplot}{Define if a boxplot has to be overdrawn (set to TRUE when the number of points is big enough)}
}
\description{
This function allows plottin a dotplot and eventually overplotting a boxplot on it
}
\examples{
expres <-c(rnorm(10,5,2), rnorm(10,10,2))
trat <- as.factor(c(rep("CT",10), rep("TR",10)))
titol <- "Treatment effect"
groupLab<- c("Control", "Treatment")
boxplotWithDots(myExpres=expres, lev=trat, aTitle=titol, groupLabels=groupLab)
}

